requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.3
flask-migrate==4.0.0
flask-login==0.6.2
pandas==2.0.1
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
Werkzeug==2.2.3
cryptography==41.0.3
email-validator==2.0.0
python-dotenv==1.0.0
secrets==1.0.2
mysqlclient==2.2.0
PyMySQL==1.1.0
